........
  r6504 | p_lindheimer | 2008-08-29 11:13:55 -0700 (Fri, 29 Aug 2008) | 1 line
  
  don't define WEBROOT if AMPWEBADDRESS is not defined, it breaks the graphing libraries (which seem to otherwise function fine without it defined)
........
  r6505 | p_lindheimer | 2008-08-29 14:55:13 -0700 (Fri, 29 Aug 2008) | 1 line
  
  make max files for call monitor configurable within main.conf.php, default remains the same
........
  r6506 | p_lindheimer | 2008-08-29 15:16:05 -0700 (Fri, 29 Aug 2008) | 1 line
  
  return the raw Asterisk version info obtained from checking the version info re #2750 and re #3128
........
  r6513 | p_lindheimer | 2008-08-30 20:53:16 -0700 (Sat, 30 Aug 2008) | 1 line
  
  re #3124 put url in auto-generated warnings to reference further help, still need to update all the fixed files with similar warnings but they should consider including some custom instructions such as sip.conf can suggest what sort of sip settings go in which includes files, etc.
........
  r6515 | p_lindheimer | 2008-08-31 07:35:02 -0700 (Sun, 31 Aug 2008) | 1 line
  
  re #3134 makes ajax refresh rates for dashboard configurable
........
  r6519 | p_lindheimer | 2008-08-31 15:40:00 -0700 (Sun, 31 Aug 2008) | 1 line
  
  re #3131 add Mixmonitor and Stopmonitor to extension class
........
  r6525 | p_lindheimer | 2008-08-31 16:36:33 -0700 (Sun, 31 Aug 2008) | 49 lines
  
  closes #3137
  
  db_connect.php:
   - removed all checks for notification panel when in quietmode since
     it's usually used for ajax calls and other light weight calls
     that can do without the overhead.
  
  config.php:
   - added restrictmods parameter, populated with '/' separted module
     names. If passed, only these modules will have their functions.inc.php
     files included. Used for light weight ajax and similar calls, to
     avoid the overhead if we know we don't need those.
  
   - removed all checks for notification panel when in quietmode. Removed
     the inclusion of hooks when in quietmode.
  
  functions.inc.php:
   - added modulelist class, used to cache the module xml structure which
     is now serialized and saved to the db to avoid the heavy xml processing
     overhead of reading all the module.xml files. (should probably move
     most of module_getinfo() into the class but for starters just did minimal
     changes to avoid introducing bugs since this is pretty sensitive).
   - fixed several instances of using " instead of ' in sql satements which
     breaks sqlite3 queries often.
   - replaced all instances of addslashes() with $db->escapeSimple() since
     sqlite3 is not compatible with addslashes() and this makes it db
     independent through pear.
   - added third argument to module_getinfo() defaulting in false. Setting to
     true forces a read of all the module.xml files. Module Admin and
     module_admin should always set this to true.
  
  header.php
   - removed all checks for notification panel when in quietmode
  page.modules.php
   - added third true paramter to module_getinfo
........
  r6526 | p_lindheimer | 2008-08-31 17:16:27 -0700 (Sun, 31 Aug 2008) | 49 lines
  
  closes #3137
  
  module_admin
   - added third true paramter to module_getinfo
........

+++ amp_conf/htdocs/admin/common/db_connect.php
@@ -71,32 +71,34 @@ if(DB::isError($db)) {
 
 // Now send or delete warning wrt to default passwords:
 //
-$nt = notifications::create($db);
+if (!$quietmode) {
+	$nt = notifications::create($db);
 
-if ($amp_conf['AMPDBPASS'] == $amp_conf_defaults['AMPDBPASS'][1]) {
-	$nt->add_warning('core', 'AMPDBPASS', _("Default SQL Password Used"), _("You are using the default SQL password that is widely known, you should set a secure password"));
-} else {
-	$nt->delete('core', 'AMPDBPASS');
-}
-
-// Check and increase php memory_limit if needed and if allowed on the system
-//
-$current_memory_limit = rtrim(ini_get('memory_limit'),'M');
-$proper_memory_limit = '100';
-if ($current_memory_limit < $proper_memory_limit) {
-	if (ini_set('memory_limit',$proper_memory_limit.'M') !== false) {
-		$nt->add_notice('core', 'MEMLIMIT', _("Memory Limit Changed"), sprintf(_("Your memory_limit, %sM, is set too low and has been increased to %sM. You may want to change this in you php.ini config file"),$current_memory_limit,$proper_memory_limit));
+	if ($amp_conf['AMPDBPASS'] == $amp_conf_defaults['AMPDBPASS'][1]) {
+		$nt->add_warning('core', 'AMPDBPASS', _("Default SQL Password Used"), _("You are using the default SQL password that is widely known, you should set a secure password"));
 	} else {
-		$nt->add_warning('core', 'MEMERR', _("Low Memory Limit"), sprintf(_("Your memory_limit, %sM, is set too low and may cause problems. FreePBX is not able to change this on your system. You should increase this to %sM in you php.ini config file"),$current_memory_limit,$proper_memory_limit));
+		$nt->delete('core', 'AMPDBPASS');
 	}
-} else {
-	$nt->delete('core', 'MEMLIMIT');
-}
 
-// send error if magic_quotes_gpc is enabled on this system as much of the code base assumes not
-//
-if(get_magic_quotes_gpc()) {
-	$nt->add_error('core', 'MQGPC', _("Magic Quotes GPC"), _("You have magic_quotes_gpc enabled in your php.ini, http or .htaccess file which will cause errors in some modules. FreePBX expects this to be off and runs under that assumption"));
-} else {
-	$nt->delete('core', 'MQGPC');
+	// Check and increase php memory_limit if needed and if allowed on the system
+	//
+	$current_memory_limit = rtrim(ini_get('memory_limit'),'M');
+	$proper_memory_limit = '100';
+	if ($current_memory_limit < $proper_memory_limit) {
+		if (ini_set('memory_limit',$proper_memory_limit.'M') !== false) {
+			$nt->add_notice('core', 'MEMLIMIT', _("Memory Limit Changed"), sprintf(_("Your memory_limit, %sM, is set too low and has been increased to %sM. You may want to change this in you php.ini config file"),$current_memory_limit,$proper_memory_limit));
+		} else {
+			$nt->add_warning('core', 'MEMERR', _("Low Memory Limit"), sprintf(_("Your memory_limit, %sM, is set too low and may cause problems. FreePBX is not able to change this on your system. You should increase this to %sM in you php.ini config file"),$current_memory_limit,$proper_memory_limit));
+		}
+	} else {
+		$nt->delete('core', 'MEMLIMIT');
+	}
+
+	// send error if magic_quotes_gpc is enabled on this system as much of the code base assumes not
+	//
+	if(get_magic_quotes_gpc()) {
+		$nt->add_error('core', 'MQGPC', _("Magic Quotes GPC"), _("You have magic_quotes_gpc enabled in your php.ini, http or .htaccess file which will cause errors in some modules. FreePBX expects this to be off and runs under that assumption"));
+	} else {
+		$nt->delete('core', 'MQGPC');
+	}
 }
